+++ customer-service/src/main/java/me/wheelchen/coffeebucks/customer/CustomRunner.java
@@ -2,16 +2,21 @@ package me.wheelchen.coffeebucks.customer;
 
 import lombok.extern.slf4j.Slf4j;
 import me.wheelchen.coffeebucks.customer.model.Coffee;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.joda.money.CurrencyUnit;
+import org.joda.money.Money;
 import org.springframework.boot.ApplicationArguments;
 import org.springframework.boot.ApplicationRunner;
+import org.springframework.core.ParameterizedTypeReference;
+import org.springframework.http.HttpMethod;
+import org.springframework.http.MediaType;
+import org.springframework.http.RequestEntity;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriComponentsBuilder;
 
-import java.math.BigDecimal;
 import java.net.URI;
+import java.util.List;
 
 /**
  * @author Kelvin Chen
@@ -29,27 +34,36 @@ public class CustomRunner implements ApplicationRunner {
 
     @Override
     public void run(ApplicationArguments args) throws Exception {
-        // === GET 请求 & 序列化===
-        // 构造URI
+        // Request - 1
+        String baseUri = "http://localhost:8080/";
         URI uri = UriComponentsBuilder
-                .fromUriString("http://localhost:8080/coffee/{id}")
-                .build(1);
-        // 发送请求
-        ResponseEntity<Coffee> c = restTemplate.getForEntity(uri, Coffee.class);
-        log.info("Response Status: {}, Response Headers: {}", c.getStatusCode(), c.getHeaders().toString());
-        log.info("Coffee: {}", c.getBody());
+                .fromUriString(baseUri + "coffee/?name={name}")
+                .build("mocha");
+        RequestEntity<Void> req = RequestEntity.get(uri)
+                //设置返回接受类型
+                .accept(MediaType.APPLICATION_XML)
+                .build();
+        ResponseEntity<String> resp = restTemplate.exchange(req, String.class);
+        log.info("Response Status: {}, Response Headers: {}", resp.getStatusCode(), resp.getHeaders().toString());
+        log.info("Coffee: {}", resp.getBody());
 
-        // === POST 请求 & 序列化 ===
-        String coffeeUri = "http://localhost:8080/coffee/";
+        // Request - 2
+        String coffeeUri = baseUri + "coffee/";
         Coffee request = Coffee.builder()
                 .name("Americano")
-                .price(BigDecimal.valueOf(25.00))
+                .price(Money.of(CurrencyUnit.of("CNY"), 25.00))
                 .build();
         Coffee response = restTemplate.postForObject(coffeeUri, request, Coffee.class);
         log.info("New Coffee: {}", response);
 
-        // === GET 请求 ===
-        String s = restTemplate.getForObject(coffeeUri, String.class);
-        log.info("String: {}", s);
+        // Request - 3 对于泛型的处理
+        ParameterizedTypeReference<List<Coffee>> ptr =
+                new ParameterizedTypeReference<List<Coffee>>() {
+        };
+
+        ResponseEntity<List<Coffee>> list = restTemplate
+                .exchange(coffeeUri, HttpMethod.GET, null, ptr);
+        list.getBody().forEach(c -> log.info("Coffee: {}", c));
+
     }
 }